add apollo plugin test separately
